requirements.txt
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.2
aiofiles>=23.1.0
python-dotenv>=1.0.0
httpx>=0.24.0
